Add interface to enable HapticGenerator for Ringtone.
HapticGenerator is an audio effect to generate haptic data from audio
data. An important usecase for HapticGenerator is ringtone. However, it
is not easy to use audio effect for the Ringtone class as the
Ringtone's lifecycle is different from AudioTrack/MediaPlayer. In that
case, add an interface in Ringtone to enable HapticGenerator. The
HapticGenerator will only be added if the effect is available for the
device.

Bug: 179080948
Test: atest RingtoneTest
Change-Id: Ia0416f6b6c6f311627be88dd9a57a989379ac947