Bubble API: suppress bubble
Apps want to hide their bubble when/if the user is viewing
the same content outside of the bubble (e.g. bubble for
alice, user opens up conversation in main app for alice).

Devs can set a flag on BubbleMetadata to request this
behavior.

The match of bubble to activity is done via a locusId. So
apps must set a locusId on the notification & activity to
make this match (done in other CL).

Bug: 170267239
Test: atest NotificationManagerTest NotificationTest (see CTS CL)
Change-Id: I9f2efd3fd1ba4abdd8e579b51245d39a6be9186f