Only check canSpecifyOrientation at leaf Task
Before, we check if the task can specify orientation to determine if the
task should be letterboxed for orientatoin request, with which we are
using the activity type of the top leaf Task for its sibling Task.

Now, we only check for leaf Task, so that we won't letterbox HOME task
even if it is not the top most.

Bug: 179362640
Test: atest WmTests:TaskTests
Test: manually with split screen
Change-Id: I92e1c1e8c33bf6877ff3aeec94dc8ab1a3ba0c3e