Add IllegalStateException with packageName on null info
This change returns the behavior of initializeJavaContextClassLoader in
LoadedApk to throw an IllegalStateException calling out the package it
attempted to fetch. This was removed when client-side caching was added
resulting in a different stack trace without this critical piece of
information that could help make sense of why this is happening.

Bug: 180418767
Bug: 140788621
Test: N/A; builds
Change-Id: Ia1ceff922f693093012eb6e3e0f6e1d90222cde7