Handle multiple Activities matching eSOS intent
We should be able to handle multiple emergengy sos activities on the
device without showing a disambig dialog. This CL adds a config string
to set the package name of the preferred app to handle the emergengy sos
intent (not using RoleManager#ROLE_EMERGENCY since SysUI doesn't have
the necessary permission).

The preference is given to the activity as follows:

1. Activity is from a system app with package name matching new config string
2. Activity is from a system app
3. Gesture becomes a no-op (will not match non-system activities)

Test: manual {StatusBar makes testing a launched activity
challenging}

Bug: 181610572
Change-Id: I862b4e3edf990d5f011aa0dc073070ede8b474b1