Workaround to use System UI Context
... for SystemGesturesPointerEventListener

This CL passes System UI Context to SystemGesturesPointerEventListener
instead of System Context's DisplayContext because DisplayContext is
not a UI context and violates IncorrectContextViolation when trying
to access GestureDetector.

Note that we should still migrate the usage to window context because it
may still an issue if device can change its display metrics at runtime.
[1] attempted to do so but hit performance issue b/181814478.

[1]: 313d5a34bdf1fbb47247ff0b4f93d4b6c63e240e

Bug: 177650348
Test: atest WmTests
Change-Id: Ie7695270d2e2dd15c749c07c2bde2f6f98bc5b59