Rename #updateInputMethodWindowStatus() to #setDismissImeOnBackKeyPressed()
This is a purely mechanical method renaming.  There should be no
behavior chagne.

To summarize,

  WindowManagerInternal#updateInputMethodWindowStatus()

should be renamed to

  WindowManagerInternal#setDismissImeOnBackKeyPressed()

and this CL actually does so with also removing unused parameters.

The method in question was originally added by me for the
true-IME-target-window project [1] at Bug 34628091 then updated to
propagate one more boolean value from InputMethodManagerService to
PhoneWindowManager to address a feature request from Wear team [2] at
Bug 33824860.

After that, we also introduced another method

   WindowManagerInternal#updateInputMethodTargetWindow()

for the true-IME-target-window project [3] and that's what we ended up
having relied on.  Given that #updateInputMethodWindowStatus() is used
only for Bug 33824860 to propagate "dismissImeOnBackKeyPressed"
parameter from IMMS to PhoneWindowManager, it'd make more sense to
rename its method name as well.

 [1]: I54fb1faf513883a54293a756e4c9d2ae9453d778
      c1628ff53821788fd70b83d61a2fc6647395da76
 [2]: I20721547c73360a70b5fc5cbe06824d577d1768a
      7827603fb677c924ed3391b1c770f318f9c03289
 [3]: I032b91ce722a16b17518a5d88015c574d7d3e51b
      afb7775f053185590c7729ddd2c4420cd86a2438

Bug: 171827834
Test: presubmit
Change-Id: I3b32d43615453d94d5df1565fe1ce81dc08526de