Don't send long-press event if KEYCODE_UNKNOWN
Gates sending a long-press event on the keycode not being unknown.
This brings the behavior into line with the other times events are
sent from KeyButtonView and fixes a bug where long-pressing the
recents button in 3-button mode broke screenshots. When this
happens we send an ACTION_DOWN event, but no ACTION_UP event,
which prevents the key combination manager from correctly
recognizing keychords.

Bug: 181178545
Test: manual; repro = long-press the recents button in 3-button
mode, then attempt to take a screenshot with power+vol_down.

Change-Id: I26d232067befc3aca04c74e4ef8f4ad4ac629934